Move Font and Text definitions to "text_renderer.c"
for simplifying the Global Code pages "Fonts" and "Texts"
where users should define their new Font and Text variables.

Also refactor "fontTool.c": rename local variable asciiNum in
printText() to letterNum to better represent its usage. The
variable is used for fetching the width of a character from
the array fontCharWidths[]. The number is not an ASCII value
of a character, but an array index calculated with fromASCII()
from the ASCII value of a character.