Add setTextPosition and modify setTextParent
A new option has been added to setTextParent,
the option to keep the current position of the text.
This makes it so that the text's position relative
to the parent is calculated based on its current
position.

Otherwise the text's current coordinates
just become relative to the parent's position
and thus cause the text to teleport to a new
position when first refreshed after changing
the parent.